Implement Authentication
Implemented a very simple authentication to authenticate the users of
API. It is a token based authentication system. Upon registraion the client
is sent a token in a HTTP only cookie and the client sends the same token
with each request.

Following routes and controllers are implemented to support
authentication

- Users Controller

Paths
- /auth #handles sign_up
- /auth/sign_in
- /auth/sign_out